requirements.txt
telethon>=1.41.0
python-dotenv>=1.0.1
openai>=2.0.0
pydantic>=2.12.0
pytz>=2024.1
feedparser>=6.0.0